rename to "lcd-stm32f7.h"
